+++ src/user.c
@@ -1,8 +1,13 @@
 #include <stdlib.h>
+#include <unistd.h>
 #include <string.h>
 #include <stdio.h>
 #include <termios.h>
+#include <sys/utsname.h>
 #include "jobs.h"
+#include "term.h"
+
+#define VERSION "0"
 
 char *_runame = 0;
 char *_xuname = 0;
@@ -12,24 +17,40 @@ char *runame(void)
   return _runame;
 }
 
-static void _logout(void)
+void version(char *unused)
+{
+  struct utsname luname = { 0 };
+  char ttyname[32];
+
+  uname(&luname);
+
+  fprintf(stderr, "\r\n%s %s.%s DDT.%s.\r\n",
+	  luname.nodename,
+	  luname.sysname,
+	  luname.release,
+	  VERSION);
+  if (!ttyname_r(0, ttyname, 32))
+    fprintf(stderr, "%s\r\n", ttyname);
+}
+
+void outtest (char *ignore)
 {
   massacre(NULL);
-  free(_runame);
-  _runame = 0;
-  if (_xuname)
-    {
-      free(_xuname);
-      _xuname = 0;
-    }
 }
 
 void logout (char *ignore)
 {
   fputs("\r\n", stderr);
+
   if (_runame)
-    _logout();
+    outtest(NULL);
+
   exit (0);
+}
+
+void intest (char *ignore)
+{
+  fputs("\r\n", stderr);
 }
 
 void login_as (char *name)
@@ -43,12 +64,24 @@ void login_as (char *name)
   _xuname = strdup(name);
   clobrf = 1;
   genjfl = 1;
-  fprintf (stderr, "\r\nWelcome, %s\r\n", _xuname);
+
+  intest(NULL);
 }
 
 void chuname (char *name)
 {
   if (_runame)
-    _logout();
+    {
+      outtest(NULL);
+      free(_runame);
+      _runame = 0;
+      if (_xuname)
+	{
+	  free(_xuname);
+	  _xuname = 0;
+	}
+    }
+  clear(NULL);
+  version(NULL);
   login_as(name);
 }
