+++ src/user.h
@@ -1,5 +1,8 @@
 void logout (char *);
 void login_as (char *name);
 void chuname (char *name);
+void version(char *);
+void intest (char *);
+void outtest (char *);
 
 char *runame(void);
